miscellaneous file transfer refinement (#139)
* timeout mechanism for file transfer unhandled requests
* file transfer minor fix
* fix filentry_t release in elafile demo
* amend file transfer doc
* file transfer: allow user to send zero length data

+++ Carrier/src/filetransfer/message.h
@@ -35,7 +35,7 @@ enum {
 };
 
 typedef struct {
-    uint64_t type;
+    uint16_t type;
 } packet_t;
 
 typedef struct {
